Added 2 new classes, plus backing up
Added 2 new wrapper classes. Files and Directories. their help avoiding
potentiel issues, in case an exception happen. so it won't shut the
program down. Later we need to specify exceptions. going to use it, to
see if directories and files exist, create, replace and append data. for
now Highscore is going to use those functions.